+++ src/android-info-service.js
@@ -1,3 +1,5 @@
+/* eslint-disable promise/always-return */
+
 import appcdLogger from 'appcd-logger';
 import DetectEngine from 'appcd-detect';
 import gawk from 'gawk';
@@ -6,7 +8,7 @@ import version from './version';
 import * as androidlib from 'androidlib';
 
 import { arrayify, debounce, get, mergeDeep } from 'appcd-util';
-import { bat, cmd, exe } from 'appcd-subprocess';
+import { cmd, exe } from 'appcd-subprocess';
 import { DataServiceDispatcher } from 'appcd-dispatcher';
 import { isFile } from 'appcd-fs';
 
@@ -29,9 +31,8 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 		this.data = gawk({
 			devices: [],
 			emulators: [],
-			ndk: [],
-			sdk: [],
-			targets: {}
+			ndks: [],
+			sdks: []
 		});
 
 		/**
@@ -61,9 +62,9 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 
 		if (cfg.android) {
 			this.userADB = cfg.android.executables && cfg.android.executables.adb;
-
 			mergeDeep(androidlib.options, cfg.android);
 		}
+		gawk.watch(cfg, 'android', () => mergeDeep(androidlib.options, cfg.android || {}));
 
 		await Promise.all([
 			this.initNDKs(),
@@ -78,8 +79,6 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 	 * @access private
 	 */
 	async initNDKs() {
-		const paths = arrayify(get(this.config, 'android.ndk.searchPaths'), true).concat(androidlib.ndk.ndkLocations[process.platform]);
-
 		this.ndkDetectEngine = new DetectEngine({
 			checkDir(dir) {
 				try {
@@ -93,7 +92,10 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 			exe:      `ndk-build${cmd}`,
 			multiple: true,
 			name:     'android:ndk',
-			paths,
+			paths: [
+				...arrayify(get(this.config, 'android.ndk.searchPaths'), true),
+				...androidlib.ndk.ndkLocations[process.platform]
+			],
 			processResults: async (results, engine) => {
 				if (results.length > 1) {
 					results.sort((a, b) => version.compare(a.version, b.version));
@@ -124,11 +126,16 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 		});
 
 		// listen for ndk results
-		this.ndkDetectEngine.on('results', results => {
-			gawk.set(this.data.ndk, results);
-		});
+		this.ndkDetectEngine.on('results', results => gawk.set(this.data.ndks, results));
 
 		await this.ndkDetectEngine.start();
+
+		gawk.watch(this.config, [ 'android', 'ndk', 'searchPaths' ], value => {
+			this.ndkDetectEngine.paths = [
+				...arrayify(value, true),
+				...androidlib.ndk.ndkLocations[process.platform]
+			];
+		});
 	}
 
 	/**
@@ -139,8 +146,6 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 	 * @access private
 	 */
 	async initSDKsDevicesAndEmulators() {
-		const paths = arrayify(get(this.config, 'android.sdk.searchPaths'), true).concat(androidlib.sdk.sdkLocations[process.platform]);
-
 		this.sdkDetectEngine = new DetectEngine({
 			checkDir(dir) {
 				try {
@@ -151,20 +156,29 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 			},
 			depth:    1,
 			env:      [ 'ANDROID_SDK', 'ANDROID_SDK_ROOT' ],
-			exe:      [ `../../adb${exe}`, `../../android${bat}` ],
+			exe:      `../../adb${exe}`,
 			multiple: true,
 			name:     'android:sdk',
-			paths,
-			processResults: async (results, engine) => {
+			paths: [
+				...arrayify(get(this.config, 'android.sdk.searchPaths'), true),
+				...androidlib.sdk.sdkLocations[process.platform]
+			],
+			processResults: async (results) => {
 				// loop over all of the new sdks and set default version
 				if (results.length) {
+					const lookup = {};
 					let foundDefault = false;
+
 					for (const result of results) {
-						if (!foundDefault && (!engine.defaultPath || result.path === engine.defaultPath)) {
-							result.default = true;
+						result.default = false;
+						lookup[result.path] = result;
+					}
+
+					for (const path of this.sdkDetectEngine.paths) {
+						if (lookup[path]) {
+							lookup[path].default = true;
 							foundDefault = true;
-						} else {
-							result.default = false;
+							break;
 						}
 					}
 
@@ -179,39 +193,35 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 			redetect: true,
 			registryKeys: [
 				{
-					hive: 'HKLM',
-					key: 'SOFTWARE\\Wow6432Node\\Android SDK Tools',
-					name: 'Path'
+					key: 'HKLM\\SOFTWARE\\Wow6432Node\\Android SDK Tools',
+					value: 'Path'
 				},
 				{
-					hive: 'HKLM',
-					key: 'SOFTWARE\\Android Studio',
-					name: 'SdkPath'
+					key: 'HKLM\\SOFTWARE\\Android Studio',
+					value: 'SdkPath'
 				}
 			],
 			watch: true
 		});
 
 		// listen for sdk results
-		this.sdkDetectEngine.on('results', results => {
-			gawk.set(this.data.sdk, results);
-		});
+		this.sdkDetectEngine.on('results', results => gawk.set(this.data.sdks, results));
 
 		let initialized = false;
 
 		const rescan = debounce(async () => {
 			console.log('Rescanning Android emulators...');
-			const emus = await androidlib.emulators.getEmulators({ force: true, sdks: this.data.sdk });
+			const emus = await androidlib.emulators.getEmulators({ force: true, sdks: this.data.sdks });
 			console.log(`Found ${emus.length} emulators`);
 			gawk.set(this.data.emulators, emus);
 		});
 
 		appcd.fs.watch({
-			type: 'avd',
-			depth: 2,
-			paths: [ androidlib.avd.getAvdDir() ],
 			debounce: true,
-			handler: rescan
+			depth: 2,
+			handler: rescan,
+			paths: [ androidlib.avd.getAvdDir() ],
+			type: 'avd'
 		});
 
 		try {
@@ -220,84 +230,32 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 		} catch (e) {
 			console.warn('Unable to subscribe to Genymotion, reverting to watching the VirtualBox config');
 			appcd.fs.watch({
-				type: 'vboxconf',
+				debounce: true,
+				handler: rescan,
 				depth: 2,
 				paths: [
 					androidlib.virtualbox.virtualBoxConfigFile[process.platform]
 				],
-				debounce: true,
-				handler: rescan
+				type: 'vboxconf'
 			});
 		}
 
-		return new Promise((resolve, reject) => {
+		await new Promise((resolve, reject) => {
 			// if sdks change, then refresh the simulators and update the targets object
-			gawk.watch(this.data.sdk, async () => {
+			gawk.watch(this.data.sdks, async () => {
 				// we need to pause gawk so two events dont fire
 				this.data.__gawk__.pause();
 
 				console.log('Android SDK changed, rescanning emulators');
-				gawk.set(this.data.emulators, await androidlib.emulators.getEmulators({ force: true, sdks: this.data.sdk }));
+				gawk.set(this.data.emulators, await androidlib.emulators.getEmulators({ force: true, sdks: this.data.sdks }));
 
 				let adb = null;
-				let index = 1;
-				const targets = {};
 
-				for (const sdk of this.data.sdk) {
+				for (const sdk of this.data.sdks) {
 					if (adb === null || sdk.default) {
 						adb = sdk.platformTools.executables.adb || null;
 					}
-
-					for (const items of [ sdk.platforms, sdk.addons ]) {
-						for (const item of items) {
-							const abis = [];
-							if (item.abis) {
-								for (const type in item.abis) {
-									/* eslint-disable max-depth */
-									for (const abi of item.abis[type]) {
-										if (abis.indexOf(abi) === -1) {
-											abis.push(abi);
-										}
-									}
-								}
-							}
-
-							const info = {
-								id:          item.sdk,
-								abis:        abis,
-								skins:       item.skins,
-								name:        item.name,
-								type:        item.platform,
-								path:        item.path,
-								revision:    item.revision,
-								androidJar:  item.androidJar,
-								aidl:        item.aidl
-							};
-
-							if (item.basedOn) {
-								// This is an addon
-								info.type = 'add-on';
-								info.vendor = item.vendor;
-								info.description = item.description;
-								info.version = item.basedOn.version || parseInt(String(item.basedOn).replace(/^android-/, '')) || null;
-								info['based-on'] = {
-									'android-version': item.basedOn.version,
-									'api-level': item.basedOn.apiLevel
-								};
-							} else {
-								info.type = 'platform';
-								info['api-level'] = item.apiLevel;
-								info.sdk = item.apiLevel;
-								info.version = item.version;
-							}
-
-							targets[index++] = info;
-						}
-					}
 				}
-
-				// set the targets
-				gawk.set(this.data.targets, targets);
 
 				// if the user set their own adb, then use that instead of the default detected one
 				if (this.userADB) {
@@ -351,10 +309,18 @@ export default class AndroidInfoService extends DataServiceDispatcher {
 
 			this.sdkDetectEngine.start()
 				.then(async results => {
+					gawk.watch(this.config, [ 'android', 'sdk', 'searchPaths' ], value => {
+						this.sdkDetectEngine.paths = [
+							...arrayify(value, true),
+							...androidlib.sdk.sdkLocations[process.platform]
+						];
+					});
+
 					// if there's no results, then the gawk watch above never gets called
 					if (!initialized && results.length === 0) {
 						initialized = true;
-						gawk.set(this.data.emulators, await androidlib.emulators.getEmulators({ force: true, sdks: this.data.sdk }));
+						gawk.set(this.data.emulators, await androidlib.emulators.getEmulators({ force: true, sdks: this.data.sdks }));
+
 						resolve();
 					}
 				})